Complite all

+++ api/src/models.py
@@ -1,7 +1,7 @@
 """
     Here my models
 """
-from sqlalchemy import Column, Integer, Numeric, String, DateTime
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -17,18 +17,18 @@ class OneHundredReport(Base):
     PLAN = Column(String(length=1, collation='utf-8'))
     NUM_SC = Column(String(length=10))
     A_P = Column(String(length=1))
-    VR = Column(Numeric)
-    VV = Column(Numeric)
-    VITG = Column(Numeric)
-    ORA = Column(Numeric)
-    OVA = Column(Numeric)
-    OITGA = Column(Numeric)
-    ORP = Column(Numeric)
-    OVP = Column(Numeric)
-    OITGP = Column(Numeric)
-    IR = Column(Numeric)
-    IV = Column(Numeric)
-    IITG = Column(Integer)
+    VR = Column(BigInteger)
+    VV = Column(BigInteger)
+    VITG = Column(BigInteger)
+    ORA = Column(BigInteger)
+    OVA = Column(BigInteger)
+    OITGA = Column(BigInteger)
+    ORP = Column(BigInteger)
+    OVP = Column(BigInteger)
+    OITGP = Column(BigInteger)
+    IR = Column(BigInteger)
+    IV = Column(BigInteger)
+    IITG = Column(BigInteger)
     DT = Column(DateTime)
     PRIZ = Column(Integer)
